⚙️ Refactor playlist and library feature folders
