+++ client/src/features/library/components/LibraryFilter/LibraryFilter.tsx
@@ -1,8 +1,8 @@
-import Tabs from '../../../ui-library/Tabs/Tabs.jsx';
-import TabsList from '../../../ui-library/Tabs/TabsList.jsx';
-import Tab from '../../../ui-library/Tabs/Tab.jsx';
-import { useAppDispatch } from '../../../app/hooks.ts';
-import { filterLibraryItems } from '../librarySlice.ts';
+import Tabs from '../../../../ui-library/Tabs/Tabs.jsx';
+import TabsList from '../../../../ui-library/Tabs/TabsList.jsx';
+import Tab from '../../../../ui-library/Tabs/Tab.jsx';
+import { useAppDispatch } from '../../../../app/hooks.ts';
+import { filterLibraryItems } from '../../librarySlice.ts';
 
 const LibraryFilter = () => {
   const dispatch = useAppDispatch();
