+++ client/src/features/library/components/LibraryList/LibraryCard/CardImage.tsx
