+++ client/src/features/library/components/LibraryList/LibraryCard/LibraryCard.module.scss
