+++ client/src/features/library/components/LibrarySearch/LibrarySearch.tsx
@@ -1,8 +1,8 @@
 import styles from './LibrarySearch.module.scss';
 import { RiCloseLargeLine, RiSearchLine } from 'react-icons/ri';
 import { ChangeEvent, useState } from 'react';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
-import { searchLibraryItems } from '../librarySlice.ts';
+import { useAppDispatch, useAppSelector } from '../../../../app/hooks.ts';
+import { searchLibraryItems } from '../../librarySlice.ts';
 
 const LibrarySearch = () => {
   const { searchQuery } = useAppSelector((state) => state.library);
