+++ client/src/features/library/components/LibrarySort/LibrarySort.tsx
@@ -1,11 +1,11 @@
 import styles from './LibrarySort.module.scss';
-import Dropdown from '../../../ui-library/Dropdown/Dropdown.tsx';
+import Dropdown from '../../../../ui-library/Dropdown/Dropdown.tsx';
 import { RiCheckFill, RiListUnordered } from 'react-icons/ri';
-import DropdownTrigger from '../../../ui-library/Dropdown/DropdownTrigger.tsx';
-import DropdownList from '../../../ui-library/Dropdown/DropdownList.tsx';
-import DropdownItem from '../../../ui-library/Dropdown/DropdownItem.tsx';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
-import { sortLibraryItems } from '../librarySlice.ts';
+import DropdownTrigger from '../../../../ui-library/Dropdown/DropdownTrigger.tsx';
+import DropdownList from '../../../../ui-library/Dropdown/DropdownList.tsx';
+import DropdownItem from '../../../../ui-library/Dropdown/DropdownItem.tsx';
+import { useAppDispatch, useAppSelector } from '../../../../app/hooks.ts';
+import { sortLibraryItems } from '../../librarySlice.ts';
 
 const LibrarySort = () => {
   const sortBy = useAppSelector((state) => state.library.sortBy);
@@ -13,7 +13,7 @@ const LibrarySort = () => {
 
   const renderDropdownItem = (
     label: string,
-    sortType: 'alphabetical' | 'recentlyAdded',
+    sortType: 'alphabetical' | 'recentlyAdded'
   ) => {
     return (
       <DropdownItem
