+++ client/src/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, useState } from 'react';
 import styles from '../AppRoutes.module.scss';
 import Nav from './Nav/Nav';
-import Library from '../features/library/Library';
+import Library from '../features/library/components/Library';
 import Footer from './Footer/Footer.tsx';
 import Queue from '../features/queue/Queue.tsx';
 import { Link } from 'react-router-dom';
@@ -43,7 +43,7 @@ const Layout = ({ children }: LayoutProps) => {
           style={{
             display: 'flex',
             gap: '8px',
-            marginTop: '8px',
+            marginTop: '8px'
           }}
         >
           <Link to="/">Home</Link>
